arqia_-2/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mysqlclient==2.2.7
openai==1.65.2
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0

# Novas bibliotecas para análise de arquivos diversos
ezdxf==1.2.0
python-docx==1.1.0
pandas==2.2.2
openpyxl==3.1.2
celery==5.3.6
redis==5.0.4

# OPCIONAL (somente se for usar BIM/IFC)
ifcopenshell
